==> scATAC-seq/4-clustering/4-level_4/Cytotoxic/01-refinement_level3.Rmd <==
---
title: 'Cytotoxic: scATAC annotation level 4'
author: "Paula Soler-Vila & Ramon Massoni-Badosa"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```

# Introduction

The objective of this notebook is to refine the clustering annotation done at level 3. This refinement is the result of a manual curation carried out by specialists to remove poor quality cells, misclassified cells or clusters with very few cells.

# Pre-processing

## Load packages

```{r}
library(Seurat)
library(Signac)
library(tidyverse)
library(reshape2)
library(ggpubr)
library(harmony)
```


## Parameters
```{r}
cell_type = "Cytotoxic"

# Paths
path_to_CD4_T_obj <- str_c(
  here::here("scATAC-seq/results/R_objects/level_3/"),
  "CD4_T",
  "/",
  "CD4_T",
  "_integrated_level_3.rds",
  sep = ""
)

path_to_obj <- str_c(
  here::here("scATAC-seq/results/R_objects/level_3/"),
  cell_type,
  "/",
  cell_type,
  "_integrated_level_3.rds",
  sep = ""
)

path_to_obj_RNA <- str_c(
  here::here("scRNA-seq/3-clustering/5-level_5/"),
  cell_type,
    "/CD8_T_level_5_annotated_level_5.rds")

# Functions
source(here::here("scRNA-seq/bin/utils.R"))


# Colors
color_palette <-  c("#1CFFCE", "#90AD1C", "#C075A6", "#85660D", 
                    "#5A5156", "#AA0DFE", "#F8A19F", "#F7E1A0", 
                    "#1C8356", "#FEAF16", "#822E1C", "#C4451C", 
                    "#1CBE4F", "#325A9B", "#F6222E", "#FE00FA", 
                    "#FBE426", "#16FF32", "black", "#3283FE",
                    "#B00068", "#DEA0FD", "#B10DA1", "#E4E1E3",
                    "#90AD1C", "#FE00FA", "#85660D", "#3B00FB",
                    "#822E1C", "coral2", "#1CFFCE", "#1CBE4F",
                    "#3283FE", "#FBE426", "#F7E1A0", "#325A9B",   
                    "#2ED9FF", "#B5EFB5", "#5A5156", "#DEA0FD",
                    "#FEAF16", "#683B79", "#B10DA1", "#1C7F93", 
                    "#F8A19F", "dark orange", "#FEAF16", "#FBE426",  
                    "Brown")


path_to_save <- str_c(
  here::here("scATAC-seq/results/R_objects/level_4/"),
  cell_type,
  "/01.",
  cell_type,
  "_integrated_level_4.rds",
  sep = ""
)
```

# Load data

## Load CD4-T data
We are going to add a specific cluster from a previous CD4-T object that corresponds to naïve CD8-T cells.
```{r}
seurat_CD4 <- readRDS(path_to_CD4_T_obj)

DimPlot(seurat_CD4,
  pt.size = 0.1) 
```

```{r}
seurat_RNA <- readRDS(path_to_obj_RNA)

DimPlot(seurat_RNA,
  pt.size = 0.1) 

tonsil_RNA_annotation <- seurat_RNA@meta.data %>%
  rownames_to_column(var = "cell_barcode") %>%
  dplyr::filter(assay == "multiome") %>%
  dplyr::select("cell_barcode", "annotation_paper")

Naive_multiome_barcode <- tonsil_RNA_annotation[tonsil_RNA_annotation$annotation_paper == "Naive CD8 T",]$cell_barcode
doublets_multiome_barcode <- tonsil_RNA_annotation[tonsil_RNA_annotation$annotation_paper == "doublets",]$cell_barcode
```


### Sub-clustering CD4-T cells
We project the barcodes from multiome naive CD8-T onto the CD4-T object and verify the cluster that correspond to this group. We will extract and integrate it into the final CD8-T object.

```{r}
DimPlot(
  seurat_CD4, 
  reduction = "umap",
  cols.highlight = "darkred", 
  cols= "grey",
  cells.highlight= Naive_multiome_barcode,
  pt.size = 0.1
)

seurat_CD4 <- FindClusters(seurat_CD4, resolution = 0.05)

DimPlot(
  seurat_CD4, 
  reduction = "umap",
  pt.size = 0.1
)

selected_cells <- row.names(seurat_CD4@meta.data[seurat_CD4@meta.data$seurat_clusters == 1,])

seurat_CD8_naive <- subset(seurat_CD4, 
                           cells = unique(selected_cells))

DimPlot(seurat_CD8_naive,
  pt.size = 0.1) 
```

# Load Cytotoxic data
```{r}
seurat_Cytotoxic <- readRDS(path_to_obj)
seurat_Cytotoxic

DimPlot(seurat_Cytotoxic,
  pt.size = 0.1) 
```

# Merging naive CD8-T with CD8-T main object
```{r}
seurat <- merge(
  x = seurat_Cytotoxic,
  y = seurat_CD8_naive)

seurat
```

# Integration
```{r fig.wide=TRUE}
seurat <- seurat %>%
  RunTFIDF() %>%
  FindTopFeatures(min.cutoff = 10) %>%
  RunSVD()

DepthCor(seurat)
seurat <- RunUMAP(object = seurat, reduction = 'lsi', dims = 2:40)

DimPlot(seurat,
        group.by = "assay")

seurat <- RunHarmony(
  object = seurat,
  dims = 2:40,
  group.by.vars = 'assay',
  reduction = 'lsi',
  assay.use = 'peaks_macs',
  project.dim = FALSE,
  max.iter.harmony = 20
)

seurat <- RunUMAP(seurat, reduction = "harmony", dims = 2:20)
seurat <- FindNeighbors(seurat, reduction = "harmony", dims = 2:20)

DimPlot(seurat, 
        group.by = "assay",
        pt.size = 0.2)
```

# Checking low quality cells
```{r}
tonsil_ATAC_cell_barcode <- seurat@meta.data %>%
  rownames_to_column(var = "cell_barcode") %>%
  dplyr::filter(assay == "multiome") %>%
  dplyr::select("cell_barcode")

possible_doublets_ATAC <-setdiff(tonsil_ATAC_cell_barcode$cell_barcode,
                                 tonsil_RNA_annotation$cell_barcode)

seurat$quality_cells <- ifelse(colnames(seurat) %in% possible_doublets_ATAC, 
                               "Poor-quality", "Good-quality")
```

##  Visualize the projection of the problematic cells onto the scATAC-seq UMAP
```{r fig.wide=TRUE}
DimPlot(
  seurat, group.by = "quality_cells")
```

## Doublet cluster defined at scRNAseq level.
```{r fig.wide=TRUE}
DimPlot(
  seurat, 
  reduction = "umap",
  cols.highlight = "darkred", 
  cols= "grey",
  cells.highlight= doublets_multiome_barcode,
  pt.size = 0.1
)
```

## Scrublet prediction
```{r}
DimPlot(seurat, group.by = "scrublet_predicted_doublet_atac")
table(seurat$scrublet_predicted_doublet_atac)
```

## QC metrics

```{r}
qc_vars <- c(
  "nCount_peaks",
  "nFeature_peaks",
  "nucleosome_signal",
  "TSS.enrichment"
)
qc_gg <- purrr::map(qc_vars, function(x) {
  p <- FeaturePlot(seurat, features = x, max.cutoff = "q95")
  p
})
qc_gg
```

## Chromatin Signature

```{r}
qc_vars <- c("NBC.MBC", "GCBC", "PC", "CD4.T", "Cytotoxic")
qc_gg <- purrr::map(qc_vars, function(x) {
  p <- FeaturePlot(seurat, feature = x, 
                   max.cutoff = 4, min.cutoff = -4) + 
    scale_color_viridis_c(option = "magma")
  p
})
qc_gg
```

# Exclude problematic clusters and poor quality cells.

```{r}
seurat <- FindClusters(seurat, resolution = 0.2)

DimPlot(seurat, 
        label = T,
        pt.size = 0.4)
```

```{r}
selected_cells <- colnames(seurat)[!(seurat$peaks_macs_snn_res.0.2 %in% c("3","6","7"))]
selected_cells <- setdiff(selected_cells,c(doublets_multiome_barcode))

seurat_clean <- subset(seurat, 
                 cells = selected_cells, 
                 quality_cells == "Good-quality")

DimPlot(seurat_clean, 
        pt.size = 0.4)
```

# Integration
```{r fig.wide=TRUE}
seurat_clean <- seurat_clean %>%
  RunTFIDF() %>%
  FindTopFeatures(min.cutoff = 10) %>%
  RunSVD()

DepthCor(seurat_clean)
seurat_clean <- RunUMAP(object = seurat_clean, reduction = 'lsi', dims = 2:40)

DimPlot(seurat_clean,
        group.by = "assay")

seurat_clean <- RunHarmony(
  object = seurat_clean,
  dims = 2:40,
  group.by.vars = 'assay',
  reduction = 'lsi',
  assay.use = 'peaks_macs',
  project.dim = FALSE,
  max.iter.harmony = 20
)

seurat_clean <- RunUMAP(seurat_clean, reduction = "harmony", dims = 2:14)

DimPlot(seurat_clean, 
        group.by = "assay",
        pt.size = 0.5)
```


## UMAP level 1
```{r}
umap_level_1 <- FeatureScatter(
  seurat,
  "UMAP_1_level_1",
  "UMAP_2_level_1",
  group.by = "annotation_level_1"
)
umap_level_1 <- umap_level_1 +
  theme(
    #legend.position = "none",
    plot.title = element_blank()
  )
umap_level_1
```

# Save
```{r}
saveRDS(seurat_clean, path_to_save)
```

# Session Information

```{r}
sessionInfo()
```
